[core] Replaced core with index and utils
